+++ types/gui/web-components/form/WeatherStatusWebComponent.d.ts
@@ -4,6 +4,9 @@ export declare class WeatherStatusWebComponent extends AbstractStateSubscriberWe
     private elements;
     private initialize;
     connectedCallback(): void;
+    disconnectedCallback(): void;
+    private handleFlightCategoryUsChange;
+    private handleFlightCategoryIcaoChange;
     static registerElement(): void;
 }
 //# sourceMappingURL=WeatherStatusWebComponent.d.ts.map